+++ pages/form-builder/preview/settings/[[...params]].tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "next-i18next";
 import { PageProps } from "@lib/types";
 import { getServerSideProps } from "../../index";
 import { NextPageWithLayout } from "../../../_app";
-import { Settings, PreviewNavigation, Template, PageTemplate } from "@formbuilder/layout";
+import { Settings, PreviewNavigation, Template, PageTemplate } from "@components/form-builder/app";
 
 const Page: NextPageWithLayout<PageProps> = () => {
   const { t } = useTranslation("form-builder");
