+++ pages/form-builder/preview/test-data-delivery.tsx
@@ -3,7 +3,12 @@ import { useTranslation } from "next-i18next";
 import { NextPageWithLayout } from "../../_app";
 import { PageProps } from "@lib/types";
 import { getServerSideProps } from "../index";
-import { PreviewNavigation, TestDataDelivery, PageTemplate, Template } from "@formbuilder/layout";
+import {
+  PreviewNavigation,
+  TestDataDelivery,
+  PageTemplate,
+  Template,
+} from "@components/form-builder/app";
 
 const Page: NextPageWithLayout<PageProps> = () => {
   const { t } = useTranslation("form-builder");
