+++ pages/form-builder/publish.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "next-i18next";
 import { NextPageWithLayout } from "../_app";
 import { PageProps } from "@lib/types";
 import { getServerSideProps } from "./index";
-import { Publish, Template, PageTemplate } from "@formbuilder/layout/";
+import { Publish, Template, PageTemplate } from "@components/form-builder/app";
 
 const Page: NextPageWithLayout<PageProps> = () => {
   const { t } = useTranslation("form-builder");
