+++ pages/form-builder/published.tsx
@@ -3,7 +3,7 @@ import { useTranslation } from "next-i18next";
 import { NextPageWithLayout } from "../_app";
 import { PageProps } from "@lib/types";
 import { getServerSideProps } from "./index";
-import { Published, Template, PageTemplate } from "@formbuilder/layout/";
+import { Published, Template, PageTemplate } from "@components/form-builder/app";
 import { useTemplateStore } from "@formbuilder/store/useTemplateStore";
 
 const Page: NextPageWithLayout<PageProps> = () => {
